move Docker file to scaffold/dockerfile and update dockerfile generation in main
