+++ robyn/__main__.py
@@ -60,43 +60,25 @@ def create_robyn_app():
             with open(schema_path, "w") as f:
                 f.write(open("robyn/scaffold/schema.prisma", "r").read())
         else:
-            f.write(
-                """
-from robyn import Robyn
-
-app = Robyn(__file__)
-
-@app.get("/")
-def index():
-    return "Hello World!"
-
-
-if __name__ == "__main__":
-    app.start()
-            """
-            )
+            f.write(open("robyn/scaffold/no-db.py", "r").read())
 
     # Dockerfile configuration
     if docker == "Y":
         print(f"Generating docker configuration for {project_dir}")
         dockerfile_path = os.path.join(project_dir, "Dockerfile")
         with open(dockerfile_path, "w") as f:
-            f.write(
-                """
-FROM ubuntu:22.04
-
-WORKDIR /workspace
-
-RUN apt-get update -y && apt-get install -y python 3.10 python3-pip
-RUN pip install --no-cache-dir --upgrade robyn
-
-COPY ./src/workspace/
-
-EXPOSE 8080
-
-CMD ["python3.10", "/workspace/foo/app.py", "--log-level=DEBUG"]
-                """
-            )
+            if project_type == "sqlite":
+                f.write(open("robyn/scaffold/docker/Dockerfile-sqlite", "r").read())
+            elif project_type == "postgres":
+                f.write(open("robyn/scaffold/docker/Dockerfile-postgres", "r").read())
+            elif project_type == "mongo":
+                f.write(open("robyn/scaffold/docker/Dockerfile-mongo", "r").read())
+            elif project_type == "sqlalchemy":
+                f.write(open("robyn/scaffold/docker/Dockerfile-sqlalchemy", "r").read())
+            elif project_type == "prisma":
+                f.write(open("robyn/scaffold/docker/Dockerfile-prisma", "r").read())
+            else:
+                f.write(open("robyn/scaffold/docker/Dockerfile", "r").read())
     elif docker == "N":
         print("Docker not included")
     else:
