+++ robyn/scaffold/docker/Dockerfile-sqlite
@@ -1,0 +1,14 @@
+FROM ubuntu:22.04
+
+WORKDIR /workspace
+
+RUN apt-get update -y && apt-get install -y python 3.10 python3-pip
+RUN pip install --no-cache-dir --upgrade robyn
+RUN pip install sqlite3
+RUN pip install uvloop
+
+COPY . .
+
+EXPOSE 8080
+
+CMD ["python3.10", "app.py", "--log-level=DEBUG"]
