+++ robyn/scaffold/no-db.py
@@ -1,0 +1,12 @@
+from robyn import Robyn
+
+app = Robyn(__file__)
+
+
+@app.get("/")
+def index():
+    return "Hello World!"
+
+
+if __name__ == "__main__":
+    app.start()
